+++ source/driver/si2c/si2c.h
@@ -25,9 +25,9 @@ si2c_status_t si2c_trans_end(void);
 uint8_t si2c_request_from(uint8_t address, uint8_t size);
 void si2c_write_byte(uint8_t data);
 void si2c_write_bytes(uint8_t *data, uint8_t size);
-int si2c_available(void);
+uint8_t si2c_available(void);
 uint8_t si2c_peek(void);
-int si2c_read_byte(void);
-int si2c_read_bytes(uint8_t *buf, uint8_t size);
+uint8_t si2c_read_byte(void);
+uint8_t si2c_read_bytes(uint8_t *buf, uint8_t size);
 
 #endif
